no weird casting here and i didn't want to bring in gsl library

+++ Source/game.cpp
@@ -236,7 +236,7 @@ void Game::CheckAnyCollisions() {
 void Game::HandleAlienProjectile() noexcept {
 	if (++shootTimer >= shootInterval && !Aliens.empty()) {
 		shootTimer = 0;
-		const int idx = GetRandomValue(0, static_cast<int>(Aliens.size()) - 1);
+		const size_t idx = static_cast<size_t>(GetRandomValue(0, static_cast<int>(Aliens.size()) - 1));
 		const Alien& shooter = Aliens.at(idx);
 		Vector2 spawnPos = shooter.GetGunPosition(); 
 		Vector2 vel = { 0, 6 };
